Refactor Supabase client integration and update authentication handling
- Updated Supabase client creation to utilize a new client structure, allowing for token management through Clerk's session.
- Replaced direct imports of createClient from the Supabase library with imports from a new client file for better organization.
- Enhanced session management by integrating Clerk's useSession hook across multiple pages, ensuring proper token retrieval for authenticated requests.
- Removed obsolete auth-config and supabase files to streamline the codebase.
