+++ lib/supabase/client.ts
@@ -1,0 +1,13 @@
+import { createBrowserClient } from "@supabase/ssr";
+
+export function createClient(token: Promise<string | null>) {
+  return createBrowserClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      async accessToken() {
+        return await token;
+      },
+    }
+  );
+}
